[modify] 댓글 fromComment 함수 서비스단으로 이동
