+++ src/main/kotlin/kr/kro/tripsketch/dto/CommentCreateDto.kt
@@ -1,6 +1,5 @@
 package kr.kro.tripsketch.dto
 
-import kr.kro.tripsketch.domain.Comment
 import java.time.LocalDateTime
 data class CommentCreateDto(
     val tripId: String,
